[Refactor:ejr] DTO 리팩토링
