+++ src/main/java/net/fullstack7/aaa/controller/api/ItemApiController.java
@@ -1,9 +1,9 @@
 package net.fullstack7.aaa.controller.api;
 
 import lombok.RequiredArgsConstructor;
-import net.fullstack7.aaa.dto.examReqRes.ExamRequestDTO;
-import net.fullstack7.aaa.dto.examReqRes.ExamResponseDTO;
-import net.fullstack7.aaa.service.exam.ExamService;
+import net.fullstack7.aaa.common.util.ApiResponseUtil;
+import net.fullstack7.aaa.dto.exam.ExamRequestDTO;
+import net.fullstack7.aaa.service.exam.ExamServiceImpl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -11,27 +11,17 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/exam")
 @RequiredArgsConstructor
 public class ItemApiController {
-    private final ExamService examService;
+    private final ExamServiceImpl examService;
 
     @PostMapping
-    public ResponseEntity<ExamResponseDTO> handleExamRequest(@RequestBody ExamRequestDTO request) {
+    public ResponseEntity<ApiResponseUtil<?>> handleExamRequest(@RequestBody ExamRequestDTO request) {
         try {
             examService.saveExamAndItems(request);
 
-            ExamResponseDTO response = ExamResponseDTO.builder()
-                    .success(true)
-                    .message("성공ㅋ")
-                    .build();
-
-            return ResponseEntity.ok(response);
+            return ResponseEntity.ok(ApiResponseUtil.success("성공"));
             
         } catch (Exception e) {
-            ExamResponseDTO errorResponse = ExamResponseDTO.builder()
-                    .success(false)
-                    .message("님들 실패함 ㅋ: " + e.getMessage())
-                    .build();
-
-            return ResponseEntity.badRequest().body(errorResponse);
+            return ResponseEntity.internalServerError().body(ApiResponseUtil.error("에러 : " + e.getMessage()));
         }
     }
 }
