+++ src/main/java/net/fullstack7/aaa/service/exam/ExamServiceIf.java
@@ -1,0 +1,7 @@
+package net.fullstack7.aaa.service.exam;
+
+import net.fullstack7.aaa.dto.exam.ExamRequestDTO;
+
+public interface ExamServiceIf {
+  public void saveExamAndItems(ExamRequestDTO request);
+}
